実装完了 (#12)
* 作成ユーザとログイン中のユーザが一致するときのみセットを表示

* roleが先生の時のみ生徒の問題の正解率を表示表示（未検証）

* 教員としてユーザ作成ができるように変更

* Usersにteacher_idの追加

* 各セットの総学習回数の表示の追加

* 教員一覧機能の実装

* 教員詳細ページの実装

* 教員のセットを削除・編集できないように実装

* 間違えて作成した/migrationをgitignoreに追加

* 教員の問題セットも見れるようにした

* セット一覧から編集・開始できるように実装

* 教員の問題セットが見れるようにした

* each_role_pageとのマージ

* 教員の問題セットの表示

* 表示の修正

* learn_countの修正

* ちょっとした変更

* sets/<sets.id>で生徒の回答の表示機能の追加

* set_listのhtmlのバグ修正

* 学習回数の表示の追加

* Update .gitignore

---------

+++ app/src/sets/views.py
@@ -1,8 +1,9 @@
-from flask import Blueprint, render_template, redirect, request, url_for, abort,flash
+from flask import Blueprint, render_template, redirect, request, url_for, abort,flash,jsonify
 from flask_login import current_user
+from sqlalchemy.sql import text
 
 from app import app
-from app.models import db, QuestionSet,Question
+from app.models import db, User, QuestionSet,Question,UserAnswer
 
 sets = Blueprint("sets",__name__)
 
@@ -15,39 +16,47 @@ def set_list():
     db.session.commit()
     return redirect(url_for('sets.set_list'))
   else: # GET
-    set = QuestionSet.query.all()
-    return render_template("sets/sets_list.html",sets=set)
+    student_sets = QuestionSet.query.filter(QuestionSet.user_id == current_user.id).all()
+    teachers = User.query.filter(User.role==1).all()
+    teacher_sets = QuestionSet.query.filter(QuestionSet.user_id == teachers[0].id).all()
+    return render_template("sets/sets_list.html", t_sets=teacher_sets ,s_sets=student_sets)
 
 @sets.route("/sets/<int:set_id>", methods=["GET"])
 def set_detail(set_id):
-  set = QuestionSet.query.get(set_id)
-  if set is None:
-    abort(404)
+    set = QuestionSet.query.get(set_id)
+    if set is None:
+        abort(404)
 
-  return render_template("sets/sets_detail.html", set=set)
+    answers = [question.correctans for question in set.questions]
+    return render_template("sets/sets_detail.html", set=set, answers=answers)
 
 
 @sets.route("/sets/<int:set_id>/edit", methods=["GET", "POST"])
 def set_edit(set_id):
   set = QuestionSet.query.get(set_id)
-  if request.method == "POST":
-    action = request.form.get("action", "update_set")
-    if action == "update_set":
-      set.name = request.form["set_name"]
-      db.session.commit()
-      return redirect(url_for('sets.set_edit', set_id=set_id))
-    elif action == "add_question":
-      question = Question(
-        sentence=request.form.get("question_sentence", ""),
-        choice1=request.form.get("choice1", ""),
-        choice2=request.form.get("choice2", ""),
-        choice3=request.form.get("choice3", ""),
-        choice4=request.form.get("choice4", ""),
-        correctans=request.form.get("correctans", ""),
-        questionset_id=set_id)
-      db.session.add(question)
-      db.session.commit()
-  return render_template("sets/sets_edit.html", set=set)
+  if set.user_id != current_user.id:
+     flash("他人のセットは編集できません")
+     return redirect(url_for('sets.set_detail', set_id=set_id))
+  else:
+    if request.method == "POST":
+      action = request.form.get("action", "update_set")
+      if action == "update_set":
+        set.name = request.form["set_name"]
+        db.session.commit()
+        return redirect(url_for('sets.set_edit', set_id=set_id))
+      #問題の追加
+      elif action == "add_question":
+        question = Question(
+          sentence=request.form.get("question_sentence", ""),
+          choice1=request.form.get("choice1", ""),
+          choice2=request.form.get("choice2", ""),
+          choice3=request.form.get("choice3", ""),
+          choice4=request.form.get("choice4", ""),
+          correctans=request.form.get("correctans", ""),
+          questionset_id=set_id)
+        db.session.add(question)
+        db.session.commit()
+    return render_template("sets/sets_edit.html", set=set)
 
 @sets.route('/question/<int:question_id>/delete', methods=['POST'])
 def delete_question(question_id):
@@ -63,21 +72,28 @@ def delete_question(question_id):
 @sets.route('/sets/<int:set_id>/delete', methods=['POST'])
 def delete_set(set_id):
     set = QuestionSet.query.get(set_id)
-    if set is None:
-        flash('Set not found')
-        return redirect(url_for('sets.set_list'))
-    db.session.delete(set)
-    db.session.commit()
-    flash('セットを削除しました')
-    return redirect(url_for('sets.set_list'))
+    if set.user_id != current_user.id:
+       flash('他人のセットは削除できません')
+       return redirect(url_for('sets.set_list'))
+    else:
+      if set is None:
+          flash('Set not found')
+          return redirect(url_for('sets.set_list'))
+      db.session.delete(set)
+      db.session.commit()
+      flash('セットを削除しました')
+      return redirect(url_for('sets.set_list'))
 
+#問題セットの学習開始
 @sets.route("/sets/<int:set_id>/start", methods=["GET"])
 def set_start(set_id):
   set = QuestionSet.query.get(set_id)
+  set.learn_count = 0
   if set is None:
     abort(404)
-
-  # Get the questions of the set
+  set.learn_count += 1
+  db.session.commit()
+  
   questions = Question.query.filter_by(questionset_id=set.id).all()
 
   # Convert each question to a dict that can be serialized to JSON
@@ -94,5 +110,18 @@ def set_start(set_id):
       'questionset_id': question.questionset_id
     }
     questions_json.append(question_dict)
-
+  
   return render_template("sets/sets_start.html", set=set,questions=questions_json)
+
+@sets.route('/sets/submit_answer', methods=['POST'])
+def submit_answer():
+    selected_answer = request.form.get('chosenAnswer')
+    question_id = request.form.get('questionId')
+    questionset_id = request.form.get('questionSetId')
+    user_id = current_user.get_id()
+
+    answer = UserAnswer(user_id=user_id, questionset_id=questionset_id, question_id=question_id, selected_answer=selected_answer)
+    db.session.add(answer)
+    db.session.commit()
+
+    return jsonify({'success': True})
